+++ motion-prime/sections/form/password_field_section.rb
@@ -8,5 +8,13 @@ module MotionPrime
     end
     element :error_message, type: :error_message, text: proc { all_errors.join("\n") if observing_errors? }
     after_render :bind_text_input
+
+    def value
+      view(:input).text
+    end
+
+    def input?
+      true
+    end
   end
 end