+++ motion-prime/sections/form/text_field_section.rb
@@ -9,5 +9,13 @@ module MotionPrime
 
     element :error_message, type: :error_message, text: proc { observing_errors? and all_errors.join("\n") }
     after_render :bind_text_input
+
+    def value
+      view(:input).text
+    end
+
+    def input?
+      true
+    end
   end
 end