+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 === 0.8.11.beta
 * Ability to send block to after_render/before_render
+* Added ability to get all form field values via `field_values`
 
 === 0.8.10
 * Added font_name/font_size options support.